site/requirements.txt
Brotli==1.0.7
click==7.1.2
Flask==1.1.2
Flask-Compress==1.5.0
Flask-JSON==0.3.4
Flask-Limiter==1.3.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
flask-talisman==0.7.0
Flask-WTF==0.14.3
importlib-metadata==1.7.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
MarkupSafe==1.1.1
python-dotenv==0.13.0
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
